+++ util/currency_util.go
@@ -5,9 +5,9 @@ import (
 	"strconv"
 )
 
-func ConvertCentsToDollars(cents int64) float64 {
+func ConvertCentsToDollars(cents float64) float64 {
 
-	dollars := float64(cents) / 100.0
+	dollars := cents / 100.0
 	rounded := fmt.Sprintf("%.2f", dollars)
 	parsedDollar, err := strconv.ParseFloat(rounded, 64)
 	if err != nil {
